remove unneeded code
